fix: 카테고리 삭제 응답 타입 수정 (#189)
카테고리 이동 응답 타입을 서비스 간 동일하게 설정합니다.

+++ backend/category-service/src/main/java/com/example/categoryservice/infrastructure/TransactionServiceClient.java
@@ -9,5 +9,5 @@ import org.springframework.web.bind.annotation.RequestParam;
 public interface TransactionServiceClient {
 
     @PutMapping("/api/transactions/category/move")
-    ApiResponse<Long> moveCategory(@RequestParam Long oldCategoryId, @RequestParam Long newCategoryId);
+    ApiResponse<Integer> moveCategory(@RequestParam Long oldCategoryId, @RequestParam Long newCategoryId);
 }
